pyproject.toml
[project]
name = "screenshort"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mss>=10.1.0",
    "paramiko>=4.0.0",
    "pyinstaller>=6.19.0",
    "python-dotenv>=1.2.2",
    "schedule>=1.2.2",
    "scp>=0.15.0",
]
